refactor: types 폴더 생성 및 types에서 products 타입 분리
