+++ src/types/types.ts
@@ -1,14 +1,3 @@
-export type Product = {
-  id: number;
-  brand: string;
-  name: string;
-  price: number;
-  originalPrice?: number;
-  discount?: number;
-  category: string;
-  image: string;
-};
-
 export type Message = {
   id: number;
   text: string;
